Renamed Response => Protocol
