+++ lib/Extension/LanguageServer/Protocol/Range.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Phpactor\Extension\LanguageServer\Response;
+namespace Phpactor\Extension\LanguageServer\Protocol;
+
+use Phpactor\Extension\LanguageServer\Protocol\Position;
 
 class Range
 {
