+++ lib/Extension/LanguageServer/Server/Method/CompletionMethod.php
@@ -6,9 +6,9 @@ use Phpactor\CodeBuilder\Adapter\TolerantParser\TextEdit;
 use Phpactor\Completion\Core\Completor;
 use Phpactor\Completion\Core\Suggestion;
 use Phpactor\Extension\Completion\Application\Complete;
-use Phpactor\Extension\LanguageServer\Response\CompletionItem;
-use Phpactor\Extension\LanguageServer\Response\CompletionList;
-use Phpactor\Extension\LanguageServer\Response\Range;
+use Phpactor\Extension\LanguageServer\Protocol\CompletionItem;
+use Phpactor\Extension\LanguageServer\Protocol\CompletionList;
+use Phpactor\Extension\LanguageServer\Protocol\Range;
 use Phpactor\Extension\LanguageServer\Server\Method;
 use Phpactor\MapResolver\Resolver;
 use Phpactor\WorseReflection\Core\Position;
